First code commit
This script seems to run fine on my computer. I've already had it
transcribe two folders of audio successfully -- with both the number of
input files changing and output files being deleted.

The code is ugly, but I thought I would share it with you all anyways <3
Perhaps at some point the ugliness will be unbearable, but for now here
we are.